Changes manner of requiring tzdata package.
Instead of attempting to dynamic-require the tzdata package,
we search for a variable in its package info.rkt. Based on
that, we define runtime paths for the data files to allow
`raco exe` to pull them in, and we modify the zoneinfo search
path here, instead of in the tzdata package.