Changes to system time zone detection:
1. Instead of trying to use the results of the first test that
returns a string, we now accumulate all of the string results
in a list and take the first one that is a real IANA time zone
id.

2. On UNIX systems, the /etc/localtime test has been changed.
   a. The actual location of the zoneinfo directory is not useful
for this test, because it is allowed to be in a location that
the operating system does not know about. For example, if you
use the tzdata package, then tzinfo will make use of its
zoneinfo files, which do not live in the standard system
location. But to detect the system's time zone, we need to
make use of the system's own files.

   b. We recently encountered a case where /etc/localtime
was symlinked to a path in the system's zoneinfo directory, which
was, in turn, a symlink to a different file. The method we were
using find the path of the zoneinfo file failed, because we
followed the links all the way to the wrong file. (We wanted
the one in the middle.) So, now we simply iterate through all
possible time zone ids and test if the actual file that is
named by that id (in the zoneinfo directory) is the same as
the actual file named by /etc/localtime.

   c. We split out the case that checks for an /etc/localtime
that is a copy of (not a symlink to) a zoneinfo file. We could
have used this mechanism to handle the symlink case, too, but
this is the most expensive test, so it might be worth avoiding
when we can.